OBGM-47 Refactor ProductExcelImporter
